Move layerdata and busy handling back into LayerView
This reverts commits 17b12b904809cae093aa59cfdd2e5958978eb6e0 and 66f8a62e6c1cf376ab6fb089a8d4911fc311e4b0.

# Conflicts:
#	plugins/LayerView/LayerPass.py
#	plugins/LayerView/LayerView.py
