+++ plugins/CuraEngineBackend/ProcessSlicedLayersJob.py
@@ -163,7 +163,7 @@ class ProcessSlicedLayersJob(Job):
 
         view = Application.getInstance().getController().getActiveView()
         if view.getPluginId() == "LayerView":
-            view.getLayerPass().resetLayerData()
+            view.resetLayerData()
 
         if self._progress:
             self._progress.hide()
